Ekstrak WebView Clients to webapp/client/
